pyproject.toml
[project]
name = "mju-sso-login"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.38.0",
    "webdriver-manager>=4.0.2",
]
